Moves TemplatePathDescriptor to the DataTypes folder
